test(spf): consolidate composition type-error tests into .test-d.ts
All tests in create-composition-types.test.ts were structurally typecheck
assertions (no expect() calls; the assertion was whether @ts-expect-error
consumed). Moving them into create-composition-types.test-d.ts is the
right home — that file is run via tsgo's typecheck mode and never
executes bodies at runtime.

The two cases that previously needed `if (Math.random() < 0)` runtime
guards no longer need them — the .test-d.ts file doesn't execute, so the
guard's whole purpose evaporates.
